backend/requirements.txt
flask
flask-cors
tensorflow==2.20.0    
pillow
numpy
opencv-python-headless
gunicorn    